+++ blocks/load.js
@@ -1,3 +1,6 @@
 import './donation-form/index';
 import './donation-form-grid/index';
 import './donor-wall/index';
+
+// Import milestone block from Milestones src
+import '../src/Milestones/resources/js/index';